refactor: :recycle: refactor routes.tsx
