+++ src/pages/Session/components/FilterSection.tsx
