+++ src/pages/Session/components/ScheduleTableSection.tsx
